Refactor old unit tests for organization admins tests
