+++ kpi/tests/kpi_test_case.py
@@ -1,24 +1,16 @@
-# coding: utf-8
 """
 Created on Apr 6, 2015
 
 @author: esmail
 """
-import re
-
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Permission
 from django.urls import reverse
 from rest_framework import status
 
 from kpi.constants import ASSET_TYPE_COLLECTION
-
-from ..models.asset import Asset
-from ..models.object_permission import ObjectPermission
-
 # FIXME: Remove the following line when the permissions API is in place.
 from .base_test_case import BaseTestCase
 from .test_permissions import BasePermissionsTestCase
+from ..models.asset import Asset
 
 
 class KpiTestCase(BaseTestCase, BasePermissionsTestCase):
@@ -39,55 +31,6 @@ class KpiTestCase(BaseTestCase, BasePermissionsTestCase):
         if username and password:
             kwargs = {'username': username, 'password': password}
         self.assertEqual(self.client.login(**kwargs), expect_success)
-
-    @staticmethod
-    def _url_to_uid(url):
-        return re.match(r'.+/(.+)/.*$', url).groups()[0]
-
-    def url_to_obj(self, url):
-        uid = self._url_to_uid(url)
-        if uid.startswith('a'):
-            klass = Asset
-        elif uid.startswith('p'):
-            klass = ObjectPermission
-        else:
-            raise NotImplementedError()
-        obj = klass.objects.get(uid=uid)
-        return obj
-
-    def obj_to_url(self, obj):
-        # Add more types as you need them
-        if isinstance(obj, ObjectPermission):
-            return reverse(
-                self._get_endpoint('asset-permission-assignment-detail'),
-                kwargs={'parent_lookup_asset': obj.asset.uid, 'uid': obj.uid},
-            )
-        if isinstance(obj, Permission):
-            return reverse(
-                self._get_endpoint('permission-detail'),
-                kwargs={'codename': obj.codename},
-            )
-        elif isinstance(obj, get_user_model()):
-            return reverse(
-                self._get_endpoint('user-kpi-detail'),
-                kwargs={'username': obj.username},
-            )
-        raise NotImplementedError
-
-    def get_asset_perm_assignment_list_url(self, asset):
-        return reverse(
-            self._get_endpoint('asset-permission-assignment-list'),
-            kwargs={'parent_lookup_asset': asset.uid}
-        )
-
-    def get_urls_for_asset_perm_assignment_objs(self, perm_assignments, asset):
-        return [
-            self.absolute_reverse(
-                self._get_endpoint('asset-permission-assignment-detail'),
-                kwargs={'uid': uid, 'parent_lookup_asset': asset.uid},
-            )
-            for uid in perm_assignments.values_list('uid', flat=True)
-        ]
 
     def create_collection(self, name, owner=None, owner_password=None,
                           **kwargs):
